+++ feature/books/data/src/commonMain/kotlin/com/retro99/books/data/model/ReadaloudLocalModel.kt
@@ -10,7 +10,7 @@ data class ReadaloudLocalModel(
     override val missing: Int?,
     override val status: String?,
     override val currentStage: String?,
-    override val stageProgress: Int?,
+    override val stageProgress: Double?,
     override val queuePosition: Int?,
     override val restartPending: Boolean?,
     override val createdAt: String?,
