+++ lib/database/api/src/commonMain/kotlin/com/retro99/database/api/books/ReadaloudEntity.kt
@@ -10,7 +10,7 @@ interface ReadaloudEntity {
     val missing: Int?
     val status: String?
     val currentStage: String?
-    val stageProgress: Int?
+    val stageProgress: Double?
     val queuePosition: Int?
     val restartPending: Boolean?
     val createdAt: String?
